+++ src/com/github/idea/arthas/plugin/action/terminal/tunnel/ArthasTerminalManager.java
@@ -1,11 +1,13 @@
-package com.github.idea.arthas.plugin.utils;
+package com.github.idea.arthas.plugin.action.terminal.tunnel;
 
-import com.github.idea.arthas.plugin.action.terminal.ClearAllAction;
-import com.github.idea.arthas.plugin.action.terminal.ModifyRerunAction;
-import com.github.idea.arthas.plugin.action.terminal.RerunAction;
-import com.github.idea.arthas.plugin.action.terminal.StopAction;
+import com.github.idea.arthas.plugin.action.terminal.tunnel.action.ClearAllAction;
+import com.github.idea.arthas.plugin.action.terminal.tunnel.action.ModifyRerunAction;
+import com.github.idea.arthas.plugin.action.terminal.tunnel.action.RerunAction;
+import com.github.idea.arthas.plugin.action.terminal.tunnel.action.StopAction;
 import com.github.idea.arthas.plugin.common.pojo.AgentInfo;
 import com.github.idea.arthas.plugin.common.pojo.TunnelServerInfo;
+import com.github.idea.arthas.plugin.utils.Icons;
+import com.github.idea.arthas.plugin.utils.StringUtils;
 import com.intellij.execution.DefaultExecutionResult;
 import com.intellij.execution.ExecutionManager;
 import com.intellij.execution.Executor;
