+++ src/com/github/idea/arthas/plugin/ui/ArthasTerminalOptionsDialog.java
@@ -4,9 +4,9 @@ import com.github.idea.arthas.plugin.common.pojo.AgentInfo;
 import com.github.idea.arthas.plugin.common.pojo.TunnelServerInfo;
 import com.github.idea.arthas.plugin.common.swing.PlaceholderTextSearchField;
 import com.github.idea.arthas.plugin.setting.AppSettingsState;
-import com.github.idea.arthas.plugin.utils.ArthasTerminalManager;
+import com.github.idea.arthas.plugin.action.terminal.tunnel.ArthasTerminalManager;
 import com.github.idea.arthas.plugin.utils.StringUtils;
-import com.github.idea.arthas.plugin.component.ArthasTunnelServerService;
+import com.github.idea.arthas.plugin.action.terminal.tunnel.service.ArthasTunnelServerService;
 import com.intellij.openapi.editor.Editor;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.ui.Messages;
